CCDM-2-msid-tool-dir: Modulize this script and make it work in correct directories
